+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added theme-color for mobile browsers
   - Bigger app bar titles
   - apk builds are now named glasstrail-vX.X.X-release.apk
+- Refactor: Renamed HistoryScreen to FeedScreen
 
 ### Fixed
 - Fix warning in web regarding missing emoji font subset